split out ssh handling to separated class
add execute command on guest machine method

use realpath, to avoid trouble when there is no symlink to core library

rename variable to be cleaner the purpose of this variable

fix PR review comments

split  testvm classes to more files

remove duplicated TEST/BOTS definition

fix avocado selenium part to use python3

add liceneses to python files and remove one bad change

revert rename of value because of conflict with timeout library
