+++ bots/image-prepare
@@ -22,12 +22,12 @@ import os
 import sys
 import subprocess
 
+from machine import testvm
+
 BOTS = os.path.abspath(os.path.dirname(__file__))
 BASE = os.path.normpath(os.path.join(BOTS, ".."))
 TEST = os.path.join(BASE, "test")
 os.environ["PATH"] = "{0}:{1}".format(os.environ.get("PATH"), BOTS)
-
-from machine import testvm
 
 networking = None
