--- bots/machine/__init__.py
@@ -1,1 +1,0 @@
-# Place holder for python module
