--- test/avocado/cdp-driver.js
@@ -1,1 +1,0 @@
-../common/cdp-driver.js
