--- test/avocado/cdp.py
@@ -1,1 +1,0 @@
-../common/cdp.py
