+++ test/avocado/run-tests
@@ -28,12 +28,11 @@ import subprocess
 # the directory structure breaks on the second run with .pyc files
 sys.dont_write_bytecode = True
 
-sys.path.append(os.path.dirname(os.path.dirname(os.path.realpath(__file__))))
-from verify import parent
-parent # pyflakes
-
+import parent
 import testvm
-
+import testlib
+import subprocess
+import time
 machine_test_dir = "/tmp/avocado_tests"
 
 # this is where the avocado results on the test machine will be stored
@@ -51,6 +50,7 @@ def prepare_avocado_tests(machine):
     """
     machine.execute(command="mkdir -p " + avocado_results_dir)
     machine.upload(["avocado"], machine_test_dir)
+    machine.upload([parent.LOCAL_COMMON_DIR, parent.LOCAL_MACHINE_DIR], machine_test_dir, relative_dir="/")
     machine.execute(command="ls -l /tmp/avocado_tests")
 
 def tap_output(machine):
@@ -87,7 +87,17 @@ def tap_output(machine):
     else:
         print("# TESTS PASSED duration: %ds" %  json_info['time'])
 
-def run_avocado_tests(machine, avocado_tests, print_failed=True, env={}):
+def ensure_deps_installed(machine):
+    try:
+        machine.execute("python3 -m avocado --help", quiet=True)
+    except subprocess.CalledProcessError:
+        machine.execute("pip3 install avocado-framework")
+    try:
+        machine.execute("python3 -c 'import selenium'", quiet=True)
+    except subprocess.CalledProcessError:
+        machine.execute("pip3 install selenium")
+
+def run_avocado_tests(machine, avocado_tests, print_failed=True, env=[]):
     """ Execute avocado on the machine with the passed environment variables and run
         the specified tests. For example:
         run_avocado(machine,
@@ -98,7 +108,8 @@ def run_avocado_tests(machine, avocado_tests, print_failed=True, env={}):
         If 'print_failed' is True, attempt to print a list of failed tests
     """
     return_state = True
-    cmd_parts = ["python3 -m avocado run",
+    ensure_deps_installed(machine)
+    cmd_parts = env + ["python3 -m avocado run",
                  "--job-results-dir " + avocado_results_dir,
                  ' '.join([machine_test_dir + os.sep + x for x in avocado_tests]),
                  ">&2"
@@ -223,11 +234,7 @@ def run_avocado(avocado_tests, verbose, browser, download_logs, sit):
 
         if not success and sit:
             sys.stderr.write(machine.diagnose() + "\n")
-            try:
-                input = raw_input
-            except NameError:
-                pass
-            input ("Press RET to continue... ")
+            testlib.sit()
 
         return success
     finally:
