--- test/avocado/sizzle.js
@@ -1,1 +1,0 @@
-../../node_modules/sizzle/dist/sizzle.js
