--- test/avocado/tap.py
@@ -1,1 +1,0 @@
-../common/tap.py
