--- test/avocado/test-functions.js
@@ -1,1 +1,0 @@
-../common/test-functions.js
