--- test/avocado/testlib.py
@@ -1,1 +1,0 @@
-../common/testlib.py
